Move more examples and fix others
