+++ examples/circle_with_keyway.rs
@@ -1,6 +1,6 @@
 //! This example demos creating a circle with keyway shape
 
-use csgrs::{CSG, Plane};
+use csgrs::CSG;
 use std::{fs, path::Path};
 
 const PATH: &str = "stl/circle_with_keyway";
@@ -8,9 +8,6 @@ const PATH: &str = "stl/circle_with_keyway";
 fn main() {
     // Ensure the folder exists
     let _ = fs::create_dir_all(PATH);
-
-    let square_2d = CSG::square(2.0, 2.0, None); // 2x2 square, centered
-    let circle_2d = CSG::circle(1.0, 32, None);
 
     // 1. Circle with keyway
     let keyway_shape = CSG::circle_with_keyway(10.0, 64, 2.0, 3.0, None);
